move schemaSyncService
